fix(ci): full-coverage maintainability gate — split gc.py GC models into gc_models.py
src/awf/service/gc.py exceeded the 1500-line first-party limit. Extract the
pure-data GC models (WorkspaceGCCandidate/Plan/Result, the compose/worktree/
companion/claude-base callable aliases, and the terminal/protected status
frozensets) into a new gc_models.py sibling, matching the existing gc_results/
gc_classify decomposition. gc.py re-exports every moved name so
awf.service.gc.<name> stays the stable public surface for callers and tests.
